Added check for test config folder in case it already exists.
Also removed debug lines from app and apiAuth

+++ test/setup.js
@@ -4,5 +4,7 @@ var path = require('path');
 var configDestination = path.join(__dirname, '../config/site_settings/127.0.0.1:6789');
 
 after(function() {
-   fs.unlinkSync(configDestination);
+   if (!global.configExists) {
+      fs.unlinkSync(configDestination);
+   }
 });
